body_prod/manos/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PyTorch and deep learning (already included in base image)
# torch==2.1.0
# torchvision==0.16.0

# Computer vision and image processing
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# Scientific computing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3

# Machine learning
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1